feat: refactor commands and repositories, add unit tests for Client, Product, and Sale features

+++ RO.DevTest.Application/Features/Product/Commands/DeleteProductCommand/DeleteProductCommandHandler.cs
@@ -1,6 +1,7 @@
 using System.Reflection.Metadata;
 using MediatR;
 using RO.DevTest.Application.Contracts.Persistence.Repositories;
+using RO.DevTest.Domain.Exception;
 
 namespace RO.DevTest.Application.Features.Products.Commands.DeleteProductCommand
 {
@@ -10,6 +11,10 @@ namespace RO.DevTest.Application.Features.Products.Commands.DeleteProductCommand
 
     public async Task<Unit> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
     {
+      var product = await _productRepository.GetByIdAsync(request.Id);
+      if (product == null)
+        throw new BadRequestException("Produto não encontrado");
+
       await _productRepository.DeleteAsync(request.Id);
       return Unit.Value;
     }
