+++ RO.DevTest.Application/Features/Sale/Commands/DeleteSaleCommand/DeleteSaleCommandHandler.cs
@@ -3,14 +3,24 @@ using RO.DevTest.Application.Contracts.Persistence.Repositories;
 
 namespace RO.DevTest.Application.Features.Sales.Commands.DeleteSaleCommand
 {
-  public class DeleteSaleCommandHandler(ISaleRepository saleRepository) : IRequestHandler<DeleteSaleCommand, Unit>
-  {
-    private readonly ISaleRepository _saleRepository = saleRepository;
-
-   public async Task<Unit> Handle(DeleteSaleCommand request, CancellationToken cancellationToken)
+    public class DeleteSaleCommandHandler : IRequestHandler<DeleteSaleCommand, Unit>
     {
-      await _saleRepository.DeleteAsync(request.Id);
-      return Unit.Value;
+        private readonly ISaleRepository _saleRepository;
+
+        public DeleteSaleCommandHandler(ISaleRepository saleRepository)
+        {
+            _saleRepository = saleRepository;
+        }
+
+        public async Task<Unit> Handle(DeleteSaleCommand request, CancellationToken cancellationToken)
+        {
+          var sale = await _saleRepository.GetByIdAsync(request.Id);
+            if (sale == null)
+            {
+                throw new Exception("Venda não encontrada");
+            }
+            await _saleRepository.DeleteAsync(request.Id);
+            return Unit.Value; 
+        }
     }
-  }
 }
